+++ apps/readest-app/public/locales/ko/translation.json
@@ -1698,8 +1698,6 @@
   "Resume audio": "오디오 다시 따라가기",
   "Following audio": "오디오 따라가는 중",
   " · estimated": " · 추정",
-  "Nightly Builds (Unstable)": "나이틀리 빌드(불안정)",
-  "Early daily builds; may be unstable": "매일 배포되는 미리보기 빌드로, 불안정할 수 있습니다",
   "Downloading Word Wise data…": "Word Wise 데이터 다운로드 중…",
   "Word Wise": "Word Wise",
   "Show a short native-language hint above difficult words.": "어려운 단어 위에 모국어로 된 짧은 힌트를 표시합니다.",
@@ -1716,5 +1714,9 @@
   "Hint Language": "힌트 언어",
   "Data": "데이터",
   "Download data packs automatically when needed.": "필요할 때 데이터 팩을 자동으로 다운로드합니다.",
-  "Failed to check for updates": "업데이트 확인에 실패했습니다"
+  "Failed to check for updates": "업데이트 확인에 실패했습니다",
+  "Update": "업데이트",
+  "Nightly Builds": "야간 빌드",
+  "Early daily builds": "매일 제공되는 미리 보기 빌드",
+  "Privacy": "개인정보 보호"
 }
